blog/requirements.txt
asgiref==3.8.1
django-cleanup==9.0.0
docopt==0.6.2
huepy==1.2.1
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
